Abstract events nbc method, add internal txn, show tests fail, different results
