+++ eth/tracers/internal/tracetest/txnOpCodeTracer_test.go
@@ -2,6 +2,7 @@ package tracetest
 
 import (
 	"encoding/json"
+	"fmt"
 	"math/big"
 	"os"
 	"path/filepath"
@@ -41,6 +42,22 @@ func testTxnOpCodeTracer(tracerName string, dirPath string, t *testing.T) {
 		if !strings.HasSuffix(file.Name(), ".json") {
 			continue
 		}
+		// if !strings.HasSuffix(file.Name(), "erc20_transfer_netbalchanges_events.json") {
+		// 	fmt.Println("Skipping", file.Name())
+		// 	continue
+		// }
+		// if !strings.HasSuffix(file.Name(), "erc20_transfer_netbalchanges_txns.json") {
+		// 	fmt.Println("Skipping", file.Name())
+		// 	continue
+		// }
+		// if !strings.HasSuffix(file.Name(), "multi_contracts_netbalchanges_events.json") {
+		// 	fmt.Println("Skipping", file.Name())
+		// 	continue
+		// }
+		// if !strings.HasSuffix(file.Name(), "multi_contracts_netbalchanges_txns.json") {
+		// 	fmt.Println("Skipping", file.Name())
+		// 	continue
+		// }
 		file := file
 		t.Run(camel(strings.TrimSuffix(file.Name(), ".json")), func(t *testing.T) {
 			t.Parallel()
@@ -110,10 +127,10 @@ func testTxnOpCodeTracer(tracerName string, dirPath string, t *testing.T) {
 
 			if !tracesEqual(ret, test.Result) {
 				// Below are prints to show differences if we fail, can always just check against the specific test json files too!
-				// x, _ := json.MarshalIndent(ret, "  ", "  ")
+				x, _ := json.MarshalIndent(ret, "  ", "  ")
 				// y, _ := json.MarshalIndent(test.Result, "", "")
-				// fmt.Println("Trace return: ")
-				// fmt.Println(string(x))
+				fmt.Println("Trace return: ")
+				fmt.Println(string(x))
 				// fmt.Println("test.Result")
 				// fmt.Println(string(y))
 				t.Fatal("traces mismatch")
